fix: Used Response Entity on controller and removed validation and moved it to service layer. Refactored the method names.

+++ src/main/java/com/zmd/springboot/employee/dao/impl/EmployeeDAOImpl.java
@@ -40,7 +40,9 @@ public class EmployeeDAOImpl implements EmployeeDAO {
     @Override
     public void deleteById(long id) {
         Employee employee = entityManager.find(Employee.class, id);
-        entityManager.remove(employee);
+        if (employee != null) {
+            entityManager.remove(employee);
+        }
     }
 
 }
